Rework annotations to React
Change the plugin to use React instead of Angular for annotations by
reusing the existing QueryEditor and query method. The change required
a few adaptions and fixes such as the extension to support multiple
metrics per query.

While at it a few more things have been fixed or modernized. Units
are automatically set and the screenshots have been updated.